% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlsR-package.R
\docType{package}
\name{tlsR-package}
\alias{tlsR}
\alias{tlsR-package}
\title{tlsR: Detection and Spatial Analysis of Tertiary Lymphoid Structures}
\description{
Fast, reproducible detection and quantitative analysis of tertiary lymphoid structures (TLS) in multiplexed tissue imaging. Implements Independent Component Analysis Trace (ICAT) index, local Ripley's K scanning, automated K Nearest Neighbor (KNN)-based TLS detection, and T-cell clusters identification as described in Amiryousefi et al. (2025) \doi{10.1101/2025.09.21.677465}.
}
\author{
\strong{Maintainer}: Ali Amiryousefi \email{ali_amiryousefi@hms.harvard.edu} (\href{https://orcid.org/0000-0002-6317-3860}{ORCID})

Authors:
\itemize{
  \item Jeremiah Wala \email{jeremiah_wala@dfci.harvard.edu} (\href{https://orcid.org/0000-0001-6591-1620}{ORCID})
}

Other contributors:
\itemize{
  \item Peter Sorger (\href{https://orcid.org/0000-0002-3364-1838}{ORCID}) [contributor]
}

}
\keyword{internal}
